fix: clean up package architecture and improve domain type organization
- Remove broken eventsourcing-client-aggregates package
- Move Command, Event, and CommandResult schemas from protocol to store package
- Fix architectural violation in aggregates package by removing protocol dependency
- Establish clean separation between domain concepts and transport protocols
- All tests and builds pass after refactoring
