+++ packages/eventsourcing-aggregates/package.json
@@ -61,8 +61,7 @@
     "effect": "^3.17.0"
   },
   "dependencies": {
-    "@codeforbreakfast/eventsourcing-store": "workspace:*",
-    "@codeforbreakfast/eventsourcing-protocol-default": "workspace:*"
+    "@codeforbreakfast/eventsourcing-store": "workspace:*"
   },
   "devDependencies": {
     "@types/node": "24.5.2",
