+++ packages/eventsourcing-aggregates/src/lib/commandProcessingFactory.ts
@@ -1,6 +1,11 @@
 import { Effect, pipe, Stream } from 'effect';
-import { EventStoreService, beginning, toStreamId } from '@codeforbreakfast/eventsourcing-store';
-import { Command, CommandResult } from '@codeforbreakfast/eventsourcing-protocol-default';
+import {
+  EventStoreService,
+  beginning,
+  toStreamId,
+  Command,
+  CommandResult,
+} from '@codeforbreakfast/eventsourcing-store';
 import { CommandProcessingServiceInterface } from './commandProcessingService';
 import { CommandRouter } from './commandHandling';
